requirements.txt
# ── Web Framework ─────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# ── Database ───────────────────────────────────────────────────
sqlalchemy>=2.0.30
asyncpg>=0.29.0          # async PostgreSQL driver (app runtime)
psycopg2-binary>=2.9.9   # sync driver (Alembic migrations)
alembic>=1.13.0

# ── Config & Validation ────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1

# ── Testing ────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
